requirements.txt
addict==2.4.0
aioconsole==0.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiomonitor==0.7.1
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
awscli==1.38.21
backports-datetime-fromisoformat==2.0.3
backports.strenum==1.3.1
billiard==4.2.1
botocore==1.37.21
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datumaro==1.10.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
docutils==0.16
environs==14.1.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
gevent==24.11.1
geventhttpclient==2.0.2
greenlet==3.1.1
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.29.3
idna==3.10
imagesize==1.4.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.34.0
janus==2.0.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json-stream==2.3.3
json-stream-rs-tokenizer==0.4.29
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.2
langdetect==1.0.9
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.2.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openvino==2025.0.0
openvino-telemetry==2025.1.0
orjson==3.10.7
ovmsclient==2023.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pgvector==0.4.0
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyemd==1.0.0
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-rapidjson==1.20
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.3.0
redis==6.1.0
redis-tools==0.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
segment-anything==1.0
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
supervision==0.25.1
sympy==1.13.1
tabulate==0.9.0
telnetlib3==2.0.4
tensorboardX==2.6.2.2
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.1.0+cu121
torchaudio==2.1.0
torchvision==0.16.0+cu121
tornado==6.4.2
tqdm==4.67.1
trafaret==2.1.1
traitlets==5.14.3
transformers==4.50.0
triton==2.1.0
tritonclient==2.36.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yapf==0.43.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
